Change what is added to the svgfile

+++ SecondSemester/FirstProject/qry_del_m_dm_mud.c
@@ -104,6 +104,10 @@ type_mMlavltree del(type_svg SVGFILE, type_txt TXTFILE, type_mMlavltree blocks_a
     // AQUIDE verificar se apagando assim apaga também da avl, se seta pra null ou se da problema com os ponteiros
     // ajustei +-, verificar se agora ta apagando certo o block
     type_hashitem block_rect = delete_item_in_hash_table(blocks_table, formatted_cep, (void*)get_key_from_block, (void*)verify_block_found);
+    if(block_rect == NULL){
+        printf("Bloco Inexistente\n");
+        return blocks_avl;
+    }
     //type_mMlavltree block_avl = delete_item_in_mMl_avl_tree(blocks_avl, block_rect, (void*)compare_rectangles_by_x_coordinate);
     blocks_avl = delete_item_in_mMl_avl_tree(blocks_avl, block_rect, (void*)compare_rectangles_by_x_coordinate, (void*)compare_rect_blocks_cep);
 
@@ -162,8 +166,6 @@ type_mMlavltree del(type_svg SVGFILE, type_txt TXTFILE, type_mMlavltree blocks_a
 
     // DUVIDA aqui remove também as pessoas que estiverem ocupando as locações ne?
 
-    printf("aquide\n");
-    printf("reportar svg e txt\n");
     return blocks_avl;
 }
 
@@ -297,17 +299,22 @@ void _mud_svg(type_property old_property, type_property new_property, type_rect 
     double new_x = get_rect_x(new_block);
     double new_y = get_rect_y(new_block);
 
-    char color_red[8], color_white[8], color_blue[8];
+    char color_red[8], color_white[8], color_blue[8], color_black[8];
     sprintf(color_red, "red%c", '\0');
     sprintf(color_white, "white%c", '\0');
     sprintf(color_blue, "blue%c", '\0');
+    sprintf(color_black, "black%c", '\0');
 
     int old_house_number = get_property_house_number(old_property);
     int new_house_number = get_property_house_number(new_property);
+    char *old_cep = get_property_cep(old_property);
+    char *new_cep = get_property_cep(new_property);
 
     insert_line_in_svg(QDMDM_SVGFILE, old_x+old_house_number, old_y+old_house_number, new_x+new_house_number, new_y+new_house_number, color_red, 4);
     insert_circle_in_svg(QDMDM_SVGFILE, old_x+old_house_number, old_y+old_house_number, 4, color_red, color_white, 2);
     insert_circle_in_svg(QDMDM_SVGFILE, new_x+new_house_number, new_y+new_house_number, 4, color_blue, color_white, 2);
+    insert_text_in_svg(QDMDM_SVGFILE, old_x+old_house_number, old_y+old_house_number, color_black, old_cep, 8);
+    insert_text_in_svg(QDMDM_SVGFILE, new_x+new_house_number, new_y+new_house_number, color_black, new_cep, 8);
 }
 
 // A pessoa identificada por cpf muda-se para o endereço determinado pelos parâmetros.
@@ -316,7 +323,6 @@ void _mud_svg(type_property old_property, type_property new_property, type_rect 
 // Um pequeno círculo vermelho no endereço antigo, outro círculo azul no novo endereço. 
 // Ambos círculos com borda branca grossa.
 void mud(type_svg SVGFILE, type_txt TXTFILE, type_hashtable blocks_table, type_hashtable people_table, type_hashtable properties_table, char person_cpf[], char property_cep[], char cardinal_direction, int house_number, char addicional_data[]){
-    printf("estou no mud\n");
     QDMDM_SVGFILE = SVGFILE;
     QDMDM_TXTFILE = TXTFILE;
     long cpf = format_cpf(person_cpf);
@@ -330,14 +336,17 @@ void mud(type_svg SVGFILE, type_txt TXTFILE, type_hashtable blocks_table, type_h
 
     if(person == NULL) return;
 
-    long cep = format_cep_from_base36_to_base10(property_cep);
+    long new_cep_key = format_cep_from_base36_to_base10(property_cep);
     type_property new_property_ = new_property(property_cep, cardinal_direction, house_number, addicional_data);
     type_property old_property = remove_owned_property_from_person(person, 0);    
+    char* old_cep = get_property_cep(old_property);
+    char* new_cep = get_property_cep(new_property_);
+    long old_cep_key = get_property_cep_key(old_property);
 
     set_owned_properties_to_person(person, 1);
     add_persons_owned_property(person, new_property_);
     add_property_owner(new_property_, person);
-    insert_item_in_hash_table(properties_table, new_property_, cep, (void*)get_property_cep_key, (void*)compare_properties_cep);
+    insert_item_in_hash_table(properties_table, new_property_, new_cep_key, (void*)get_property_cep_key, (void*)compare_properties_cep);
 
     long(*get_key_from_block_ptr)(type_block);
     get_key_from_block_ptr = get_key_from_block;
@@ -345,11 +354,12 @@ void mud(type_svg SVGFILE, type_txt TXTFILE, type_hashtable blocks_table, type_h
     long(*verify_block_found_ptr)(type_block);
     verify_block_found_ptr = verify_block_found;
 
-    set_id(get_property_cep(old_property));
-    type_rect old_block = lookup_item_in_hash_table(blocks_table, (long)get_property_cep_key(old_property), (void*)get_key_from_block_ptr, (void*)verify_block_found_ptr);
-    set_id(get_property_cep(new_property_));
-    type_rect new_block = lookup_item_in_hash_table(blocks_table, (long)get_property_cep_key(new_property_), (void*)get_key_from_block_ptr, (void*)verify_block_found_ptr);
+    set_id(old_cep);
+    type_rect old_block = lookup_item_in_hash_table(blocks_table, old_cep_key, (void*)get_key_from_block_ptr, (void*)verify_block_found_ptr);
+    set_id(new_cep);
+    type_rect new_block = lookup_item_in_hash_table(blocks_table, new_cep_key, (void*)get_key_from_block_ptr, (void*)verify_block_found_ptr);
 
     _mud_txt(person, old_property, new_property_);
+    if(old_block == NULL || new_block == NULL ) return;
     _mud_svg(old_property, new_property_, old_block, new_block);
 }
